refactor: Image 컴포넌트 리팩터링
네이밍을 WebpImage로 수정 및 next의 Image 태그 사용
